refactor!: Move encrypt.JSONWebKey and encrypt.File to event
`encrypt` will import `event` in the future and this change prevents
circular imports.
